requirements.txt
#Core Web Framework
streamlit>=1.28.0

# PDF Processing
PyPDF2>=3.0.1

# AI/ML Libraries
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5

# Vector Search & Embeddings
faiss-cpu>=1.7.4
numpy>=1.24.0

# Text Processing & NLP
regex>=2023.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Data Handling
pandas>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Async Processing
asyncio-throttle>=1.0.2
aiohttp>=3.8.0

# Additional Utilities
typing-extensions>=4.8.0
pydantic>=2.0.0

# Development & Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0